+++ sdk/resourcemanager/cosmos/armcosmos/interfaces.go
@@ -17,24 +17,15 @@ type BackupPolicyClassification interface {
 	GetBackupPolicy() *BackupPolicy
 }
 
-// BaseCosmosDataTransferDataSourceSinkClassification provides polymorphic access to related types.
-// Call the interface's GetBaseCosmosDataTransferDataSourceSink() method to access the common type.
+// DataTransferJobPropertiesClassification provides polymorphic access to related types.
+// Call the interface's GetDataTransferJobProperties() method to access the common type.
 // Use a type switch to determine the concrete type.  The possible types are:
-// - *BaseCosmosDataTransferDataSourceSink, *CassandraDataTransferDataSourceSink, *MongoDataTransferDataSourceSink, *SQLDataTransferDataSourceSink
-type BaseCosmosDataTransferDataSourceSinkClassification interface {
-	DataTransferDataSourceSinkClassification
-	// GetBaseCosmosDataTransferDataSourceSink returns the BaseCosmosDataTransferDataSourceSink content of the underlying type.
-	GetBaseCosmosDataTransferDataSourceSink() *BaseCosmosDataTransferDataSourceSink
-}
-
-// DataTransferDataSourceSinkClassification provides polymorphic access to related types.
-// Call the interface's GetDataTransferDataSourceSink() method to access the common type.
-// Use a type switch to determine the concrete type.  The possible types are:
-// - *AzureBlobDataTransferDataSourceSink, *BaseCosmosDataTransferDataSourceSink, *CassandraDataTransferDataSourceSink, *DataTransferDataSourceSink,
-// - *MongoDataTransferDataSourceSink, *MongoVCoreDataTransferDataSourceSink, *SQLDataTransferDataSourceSink
-type DataTransferDataSourceSinkClassification interface {
-	// GetDataTransferDataSourceSink returns the DataTransferDataSourceSink content of the underlying type.
-	GetDataTransferDataSourceSink() *DataTransferDataSourceSink
+// - *DataTransferJobProperties, *SourceBlobDestinationCassandraDataTransferProperties, *SourceCassandraDestinationBlobDataTransferProperties,
+// - *SourceCassandraDestinationCassandraDataTransferProperties, *SourceMongoDestinationMongoDataTransferProperties, *SourceMongoDestinationMongoVCoreDataTransferProperties,
+// - *SourceSQLDestinationSQLDataTransferProperties
+type DataTransferJobPropertiesClassification interface {
+	// GetDataTransferJobProperties returns the DataTransferJobProperties content of the underlying type.
+	GetDataTransferJobProperties() *DataTransferJobProperties
 }
 
 // ServiceResourceCreateUpdatePropertiesClassification provides polymorphic access to related types.